+++ packages/undercut-web-2019/src/__tests__/jest.setup.js
@@ -1,0 +1,2 @@
+// Polyfills for browsers
+import "core-js/stable";