+++ packages/undercut-web-2019/src/__tests__/web.test.js
@@ -1,0 +1,3 @@
+test(`Tests should run in a browser`, () => {
+	expect(typeof navigator).toBeDefined();
+});