+++ CHANGELOG.md
@@ -9,11 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `src/utils/tests.js` isn't excluded from the package anymore, and you may use it to test your custom operations.
+- `src/utils/tests.js` is included into the package, you may use it to test your custom operations.
 
 ### Changed
 
 - **[BREAKING]** `pull`/`push`/`utils` entries where moved out of the `src` directory. So, import `undercut/push.js` instead of `undercut/src/push.js`. All entry points in the package `root` are public API, but files in the `src` are private and may change their exports, import them at your own risk.
+- **[BREAKING]** `undercut` entry point is now identical to `pull.js`, to use utils you manually need to import `undercut/utils.js`.
 
 ## [0.4.0] - 2019-10-15